+++ udata/core/dataservices/models.py
@@ -113,7 +113,10 @@ class Dataservice(WithMetrics, Owned, db.Document):
                 db.ReferenceField(Dataset),
                 nested_fields=datasets_api_fields.dataset_fields,
             )
-        )
+        ),
+        filterable={
+            'key': 'dataset',
+        },
     )
 
     @function_field(description="Link to the API endpoint for this dataservice")