+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Improve `udata db check-integrity` (perfs, Sentry notifications…) [#3026](https://github.com/opendatateam/udata/pull/3026)
 - Refactor catalog exports [#3052](https://github.com/opendatateam/udata/pull/3052)
+- Add a filter to filter dataservices by dataset [#3056](https://github.com/opendatateam/udata/pull/3056)
+- Fix reuses' datasets references [#3057](https://github.com/opendatateam/udata/pull/3057)
 - Save and show harvest logs [#3053](https://github.com/opendatateam/udata/pull/3053)
 
 ## 9.0.0 (2024-06-07)